timext: change timeout data type from unsigned int to unsigned long

+++ ports/zephyr/timext.c
@@ -7,19 +7,19 @@
 #include "libmcu/timext.h"
 #include <zephyr/kernel.h>
 
-#define time_after(goal, chasing)	((int)(goal)    - (int)(chasing) < 0)
+#define time_after(goal, chasing)	((long)(goal)    - (long)(chasing) < 0)
 
-void timeout_set(unsigned int *goal, unsigned int msec)
+void timeout_set(unsigned long *goal, unsigned long msec)
 {
 	*goal = k_uptime_get_32() + msec;
 }
 
-bool timeout_is_expired(unsigned int goal)
+bool timeout_is_expired(unsigned long goal)
 {
 	return time_after(goal, k_uptime_get_32());
 }
 
-void sleep_ms(unsigned int msec)
+void sleep_ms(unsigned long msec)
 {
 	k_sleep(K_MSEC(msec));
 }
